feat: enable fluentd logging
