+++ jina/clients/python/helper.py
@@ -5,7 +5,7 @@ import sys
 import time
 
 from ...helper import colored
-from ...logging import profile_logger
+from ...logging import default_logger
 from ...logging.profile import TimeContext
 
 
@@ -59,7 +59,7 @@ class ProgressBar(TimeContext):
         if num_bars == self.bar_len:
             sys.stdout.write('\n')
         sys.stdout.flush()
-        profile_logger.debug({'num_bars': num_bars,
+        default_logger.debug({'num_bars': num_bars,
                               'num_reqs': self.num_reqs,
                               'bar_len': self.bar_len,
                               'progress': num_bars / self.bar_len,
